fixed interface for custom base learners
It was previously not possible to pass arguments to the base learner constructor. Now the user simply initializes a base learner of the type she wants and passes that to ngboost. Ngboost clones and then fits it each time it needs to.

The alternative to this interface is to pass ngboost both the constructor and the arguments necessary to instantiate a new base learner. This is less tidy and modular, I think, but may play better with sklearn's gridsearchcv. If that turns out to be the case I will change the interface again in the future. Either way, the defaults are fine as they are and if a user is not using a custom base learner they will not notice these changes.