implemented sample weighting
Sample weighting implemented by weighting the losses where appropriate: in the construction of the base models and in the line search. The calculation of the gradients is not weighted, since it is not appropriate to do so. The gradient is simply how much the parameter should move at that point x. Weighting just means we have effectively w points at x, each of which should have the SAME gradient. That's why the construction of the base models is weighted, not the calculation of the gradient.