initial WIP
This is a major refactor of how distributions and scoring rules work together in NGBoost. The new system is more structured and organized, which should speed development and testing. This is an overview of the class inheritance:

manifold -> distribution -> Distn
|-> distributionscore ->  score  ->
 Score

For distributions, the top level class is Distn, which implements default methods for slicing and indexing. Each distribution (e.g. Normal) is a subclass of Distn which implements the parameters of that distribution as well as methods for fitting the marginal and sampling. I may eventually implement an intermediate class of,
 e.g., ContinuousDistns, which would implement the appropriate `predict()`  method, etc.

For scoring rules, it's helpful to move from granular to specific. The most specific scoring rule is a "distributionscore" (e.g. NormalLogScore), which typically implements score, d_score, and metric methods for some combination of scoring rule and distribution. Each distributionscore is a subclass of a scoringrule (e.g. LogScore). LogScore implements a default method for calculating the metric (via MC fisher) using just the d_score implementation of the more specific class. All scoringrules are a subclass of Score, which implements methods for total_score (weighted average of the score from the distributionscore) and grad (metric()^{-1}*d_score()}.

The NGBoost core code should now use manifold() to combine the user-specified scoring rule and distribution. manifold() looks up the appropriate implementation (subclass) of the scoring rule (the one for the specified distribution) and dynamically generates a new class that mixes in all of those methods with those of the distribution. The result,  which I call a "manifold" (Reimannian manifold) is a class that holds all the distributional properties, but is also endowed with methods to calculate the appropriate metrics and gradients.  The NGBoost core code will have to be slightly modified to work directly with manifold objects constructed from a score and distribution instead of with the raw score and distribution classes.

As far as the user is concerned, there is no change. pred_dist() will now return manifold objects, but since these are also distributions, the change is invisible.