pred_dist returns a strict dist, not manifold
The additional methods that come with a manifold (relative to what came with the old dist) are total_score and grad, which in the old setting were basically minor combinations/wrappers of what are now score, d_score, and metric . So relative to what we used to be returning the the user as a “dist”, the “manifold” actually doesn’t expose that much more. On the other hand, now that it’s organized like this, we don’t have to expose any of that if we don’t want and we could return what’s in dist, which is as little as the parameters (and any transformations calculated in dist.__init__), a method to fit a marginal, and a method to sample from the distribution.