Enable Base models that don't accept sample weight
This change enables using Base models that don't accept the keyword argument `sample_weight` (for example, `sklearn.neural_network.MLPRegressor`) when the user specifies or defaults to `sample_weight=None`. In this case, `fit()` is called without this keyword argument.

The alternatives would be to use the library `inspect` to check whether the base model's `fit()` accepts this keyword argument and split cases from there, or handle the exception thrown when the unexpected keyword argument is received, but this is pretty simple in comparison.